refactor(qwen3_5_moe): make the types say what the lifecycle already required
mypy went from 44 errors in these files to zero, with no change in behaviour: the example was rerun on
device and produces the same output. Two of the changes are worth more than the type checker's approval.

The KV caches were `Tensor | None`, Optional only until the runner calls bind_kv_cache. Every read site
therefore had to narrow them, forty times, for what is really a lifecycle fact. They are now private
with properties that raise and say what to call — so using a layer before its caches are bound fails
with the reason instead of with an AttributeError on None, and the type is honest at every read.

`nn.ModuleList` yields `Module`, which loses everything the layer declares, so iterating it defeated the
class-level annotations the layer now carries. A single typed accessor (`decoder_layers`) fixes every
loop and says why it exists. The alternative — casting at each of the four iteration sites — would have
been the same amount of noise mypy asked for, in four places instead of one.

The rest is annotations and three casts where an untyped helper returns Any, each with a comment naming
what makes it safe. No blanket `# type: ignore`.

Also here, the first half of the MTP head's contract: `kv_layer_spec()` and `bind_kv_cache_entry()`. The
draft layer is full attention, so it needs its own KV cache entry and its own attention metadata under a
layer index that cannot collide with the main model's. Returned as a plain tuple rather than a LayerSpec
so the module stays importable by a CPU test; the model assembles the real spec once the index is
allocated.

120 CPU tests pass; ruff and mypy are both clean on these files.

+++ vllm_neuron/model/qwen3_5_moe/vision_inputs.py
@@ -34,8 +34,8 @@ def patch_dim(vision_config) -> int:
     This is the width the checkpoint's ``patch_embed.proj`` expects, so it is derived from the config
     rather than passed in; a mismatch here would show up as a shape error deep inside the encoder.
     """
-    return (vision_config.in_channels * vision_config.temporal_patch_size
-            * vision_config.patch_size * vision_config.patch_size)
+    return int(vision_config.in_channels * vision_config.temporal_patch_size
+               * vision_config.patch_size * vision_config.patch_size)
 
 
 def block_count(total_patches: int, block_size: int) -> int:
